requirements.txt
openai
streamlit
chromadb
langchain
langchain-community
langchain-openai
langchain-google-genai
langchain-groq
python-dotenv
jupyter
ipykernel
linkedin-api
langchain_chroma